requirements.txt
django
djangorestframework
django-cors-headers
pillow
djangorestframework-simplejwt
django-filter
requests
psycopg[binary]
python-dotenv
cryptography
dj-database-url
gunicorn
whitenoise
uvicorn
